收银台 整理 优化按钮点击区域，优化登录页阴影效果，修复历史订单数据未加载完时点击打印bug，优化对话框弹出及页面切换时导航栏显示隐藏效果

+++ app/src/main/java/com/easygo/cashier/printer/local/PrinterHelpter.java
@@ -334,7 +334,7 @@ public class PrinterHelpter {
             esc.addText("  ");
             esc.addText(data.getProduct_preferential());
             esc.addText("   ");
-            esc.addText(GoodsResponse.isWeightGood(data.getType())? df_weight.format(data.getRefund_num()) + data.getG_u_symbol()
+            esc.addText(GoodsResponse.isWeightGood(data.getType())? df_weight.format(Float.valueOf(data.getRefund_num())) + data.getG_u_symbol()
                     : data.getRefund_num());
             esc.addText("  ");
             esc.addText(data.getRefund_subtotal());
